+++ Radiologo.java
@@ -3,7 +3,7 @@ public class Radiologo extends Medico {
     private String tipo;
     private double tarifa;
 
-    public Radiologo(int id, String nombre, String departamento, int experiencia, int salarioBase, String tipo, double tarifa) {
+    public Radiologo(int id, String nombre, String departamento, int experiencia, double salarioBase, String tipo, double tarifa) {
         super(id, nombre, departamento, experiencia, salarioBase);
         this.tipo = tipo;
         this.tarifa = tarifa;
